+++ javaScript/projects/weather-app/src/components/tabs/hourly/hourly.js
@@ -7,23 +7,23 @@ import createContentRows from '../../../helpers/createContentRows';
 const hourlyBuilder = {
   init(weatherData) {
     // copying specific object properties from weatherData to hourly
-    Object.keys(hourly).forEach((key) => {
-      if (!(hourly[key] instanceof Array)) {
-        Object.keys(hourly[key]).forEach((subkey) => {
-          hourly[key][subkey] = weatherData[key][subkey];
-        });
-      } else {
-        weatherData.forecast[key].forEach((day, i) => {
-          const date1 = new Date(hourly.current.last_updated);
-          hourly[key] = weatherData.forecast[key];
-          const newHours = day.hour.filter((hour) => {
-            const date2 = new Date(hour.time);
-            return date1.getTime() < date2.getTime();
-          });
-          hourly[key][i].hour = newHours;
-        });
-      }
-    });
+    // Object.keys(hourly).forEach((key) => {
+    //   if (!(hourly[key] instanceof Array)) {
+    //     Object.keys(hourly[key]).forEach((subkey) => {
+    //       hourly[key][subkey] = weatherData[key][subkey];
+    //     });
+    //   } else {
+    //     weatherData.forecast[key].forEach((day, i) => {
+    //       const date1 = new Date(hourly.current.last_updated);
+    //       hourly[key] = weatherData.forecast[key];
+    //       const newHours = day.hour.filter((hour) => {
+    //         const date2 = new Date(hour.time);
+    //         return date1.getTime() < date2.getTime();
+    //       });
+    //       hourly[key][i].hour = newHours;
+    //     });
+    //   }
+    // });
   },
   cacheDOM() {
     console.log('cacheDOM() running from hourly.js');
@@ -44,27 +44,28 @@ const hourlyBuilder = {
     const hourlyContent = createElement('section');
     hourlyContent.id = 'hourly_content';
 
-    hourly.forecastday.forEach((day) => {
-      const hourlyContentList = createElement('ol');
-      const hourlyContentHeading = createElement('h2');
-      hourlyContentList.className = 'day';
-      hourlyContentHeading.textContent = formatDate(day.date);
-      hourlyContentList.appendChild(hourlyContentHeading);
-      day.hour.forEach((hour) => {
-        hourlyContentList.appendChild(
-          createContentRows(
-            createElement,
-            [{ class: 'hour' }, { class: 'hourly_item' }],
-            formatTime(hour.time.split(' ')[1]),
-            `${Math.round(hour.temp_f)}°`,
-            hour.condition.text,
-            `${hour.chance_of_rain}%`,
-            `${hour.wind_dir} ${Math.round(hour.wind_mph)} mph`,
-          ),
-        );
-      });
-      hourlyContent.appendChild(hourlyContentList);
-    });
+    // hourly.forecastday.forEach((day) => {
+    //   const hourlyContentList = createElement('ol');
+    //   const hourlyContentHeading = createElement('h2');
+    //   hourlyContentList.className = 'day';
+    //   hourlyContentHeading.textContent = formatDate(day.date);
+    //   hourlyContentList.appendChild(hourlyContentHeading);
+    //   day.hour.forEach((hour) => {
+    //     hourlyContentList.appendChild(
+    //       createContentRows(
+    //         createElement,
+    //         [{ class: 'hour' }, { class: 'hourly_item' }],
+    //         formatTime(hour.time.split(' ')[1]),
+    //         `${Math.round(hour.temp_f)}°`,
+    //         hour.condition.icon,
+    //         hour.condition.text,
+    //         `${hour.chance_of_rain}%`,
+    //         `${hour.wind_dir} ${Math.round(hour.wind_mph)} mph`,
+    //       ),
+    //     );
+    //   });
+    //   hourlyContent.appendChild(hourlyContentList);
+    // });
 
     hourlySection.appendChild(hourlyContent);
     // temporary
@@ -73,7 +74,7 @@ const hourlyBuilder = {
 };
 
 export default function buildHourly(weatherData) {
-  hourlyBuilder.init(weatherData);
+  hourly.init('imperial', weatherData);
   return hourlyBuilder.render();
 }