+++ javaScript/projects/weather-app/src/helpers/createContentRows.js
@@ -1,18 +1,28 @@
 export default function createContentRows(fn, attributes, ...args) {
-  const containerAttributes = attributes ? attributes[0] : null;
-  const itemAttributes = attributes ? attributes[1] : null;
+  const containerAttributes = attributes ? attributes[1] : false;
+  const itemAttributes = attributes ? attributes[2] : false;
 
   const container = fn('ul');
   if (containerAttributes) {
     container.setAttributes(containerAttributes);
   }
-  args.forEach((text) => {
-    const item = fn('li');
-    if (itemAttributes) {
-      item.setAttributes(itemAttributes);
+  args.forEach((item) => {
+    // /\.(svg|png)$/
+    // console.log(item.split(/\.(svg|png)$/));
+
+    if (item) {
+      const listItem = fn('li');
+      if (itemAttributes) item.setAttributes(itemAttributes);
+      if (/\.[svg|png]{3}$/.test(item)) {
+        const img = fn('img');
+        if (/\.[svg]{3}$/.test(item)) img.setAttributes({ onload: 'SVGInject(this)' });
+        img.src = item;
+        listItem.appendChild(img);
+      } else {
+        listItem.textContent = item;
+      }
+      container.appendChild(listItem);
     }
-    item.textContent = text;
-    container.appendChild(item);
   });
   return container;
 }